Using legend.csv to verify counts in the input json. Progress on loot table 4 template.

+++ keysAndDefs.py
@@ -149,7 +149,7 @@ class LegendAsset:
         self.material = material
 
         if fixednum != '':
-            fixednum = int(fixednum)
+            fixednum = float(fixednum)
         self.fixednum = fixednum
 
         self.movable = movable.lower() in TRUES
